+++ xarray/core/computation.py
@@ -8,6 +8,8 @@ import itertools
 import operator
 import re
 
+import numpy as np
+
 from . import ops
 from .alignment import deep_align
 from .merge import expand_and_merge_variables
@@ -16,6 +18,8 @@ from .utils import is_dict_like
 
 
 _DEFAULT_FROZEN_SET = frozenset()
+_DEFAULT_FILL_VALUE = object()
+
 
 # see http://docs.scipy.org/doc/numpy/reference/c-api.generalized-ufuncs.html
 DIMENSION_NAME = r'\w+'
@@ -202,6 +206,7 @@ def apply_dataarray_ufunc(func, *args, **kwargs):
     signature = kwargs.pop('signature')
     join = kwargs.pop('join', 'inner')
     exclude_dims = kwargs.pop('exclude_dims', _DEFAULT_FROZEN_SET)
+    keep_attrs = kwargs.pop('keep_attrs', False)
     if kwargs:
         raise TypeError('apply_dataarray_ufunc() got unexpected keyword '
                         'arguments: %s' % list(kwargs))
@@ -217,12 +222,18 @@ def apply_dataarray_ufunc(func, *args, **kwargs):
     result_var = func(*data_vars)
 
     if signature.n_outputs > 1:
-        return tuple(DataArray(variable, coords, name=name, fastpath=True)
-                     for variable, coords in zip(result_var, result_coords))
+        out = tuple(DataArray(variable, coords, name=name, fastpath=True)
+                    for variable, coords in zip(result_var, result_coords))
     else:
         coords, = result_coords
-        return DataArray(result_var, coords, name=name, fastpath=True)
+        out = DataArray(result_var, coords, name=name, fastpath=True)
 
+    if keep_attrs and isinstance(args[0], DataArray):
+        if isinstance(out, tuple):
+            out = tuple(ds._copy_attrs_from(args[0]) for ds in out)
+        else:
+            out._copy_attrs_from(args[0])
+    return out
 
 def ordered_set_union(all_keys):
     # type: List[Iterable] -> Iterable
@@ -326,32 +337,53 @@ def _fast_dataset(variables, coord_variables):
 
 def apply_dataset_ufunc(func, *args, **kwargs):
     """apply_dataset_ufunc(func, *args, signature, join='inner',
-                           fill_value=None, exclude_dims=frozenset()):
+                           dataset_join='inner', fill_value=None,
+                           exclude_dims=frozenset(), keep_attrs=False):
+
+       If dataset_join != 'inner', a non-default fill_value must be supplied
+       by the user.  Otherwise a TypeError is raised.
     """
+    from .dataset import Dataset
     signature = kwargs.pop('signature')
     join = kwargs.pop('join', 'inner')
+    dataset_join = kwargs.pop('dataset_join', 'inner')
     fill_value = kwargs.pop('fill_value', None)
     exclude_dims = kwargs.pop('exclude_dims', _DEFAULT_FROZEN_SET)
+    keep_attrs = kwargs.pop('keep_attrs', False)
+    first_obj = args[0]  # we'll copy attrs from this in case keep_attrs=True
+
+    if dataset_join != 'inner' and fill_value is _DEFAULT_FILL_VALUE:
+        raise TypeError('To apply an operation to datasets with different ',
+                        'data variables, you must supply the ',
+                        'dataset_fill_value argument.')
+
     if kwargs:
         raise TypeError('apply_dataset_ufunc() got unexpected keyword '
                         'arguments: %s' % list(kwargs))
-
     if len(args) > 1:
         args = deep_align(args, join=join, copy=False, exclude=exclude_dims,
                           raise_on_invalid=False)
 
     list_of_coords = build_output_coords(args, signature, exclude_dims)
-
     args = [getattr(arg, 'data_vars', arg) for arg in args]
+
     result_vars = apply_dict_of_variables_ufunc(
-        func, *args, signature=signature, join=join, fill_value=fill_value)
+        func, *args, signature=signature, join=dataset_join,
+        fill_value=fill_value)
 
     if signature.n_outputs > 1:
-        return tuple(_fast_dataset(*args)
-                     for args in zip(result_vars, list_of_coords))
+        out = tuple(_fast_dataset(*args)
+                    for args in zip(result_vars, list_of_coords))
     else:
         coord_vars, = list_of_coords
-        return _fast_dataset(result_vars, coord_vars)
+        out = _fast_dataset(result_vars, coord_vars)
+
+    if keep_attrs and isinstance(first_obj, Dataset):
+        if isinstance(out, tuple):
+            out = tuple(ds._copy_attrs_from(first_obj) for ds in out)
+        else:
+            out._copy_attrs_from(first_obj)
+    return out
 
 
 def _iter_over_selections(obj, dim, values):
@@ -530,7 +562,8 @@ def apply_array_ufunc(func, *args, **kwargs):
 
 def apply_ufunc(func, *args, **kwargs):
     """apply_ufunc(func, *args, signature=None, join='inner',
-                   exclude_dims=frozenset(), dataset_fill_value=None,
+                   exclude_dims=frozenset(), dataset_join='inner',
+                   dataset_fill_value=_DEFAULT_FILL_VALUE, keep_attrs=False,
                    kwargs=None, dask_array='forbidden')
 
     Apply a vectorized function for unlabeled arrays to xarray objects.
@@ -581,14 +614,23 @@ def apply_ufunc(func, *args, **kwargs):
         - 'inner': use the intersection of object indexes
         - 'left': use indexes from the first object with each dimension
         - 'right': use indexes from the last object with each dimension
+    dataset_join : {'outer', 'inner', 'left', 'right'}, optional
+        Method for joining variables of Dataset objects with mismatched
+        data variables.
+        - 'outer': take variables from both Dataset objects
+        - 'inner': take only overlapped variables
+        - 'left': take only variables from the first object
+        - 'right': take only variables from the last object
+    dataset_fill_value : optional
+        Value used in place of missing variables on Dataset inputs when the
+        datasets do not share the exact same ``data_vars``. Required if
+        ``dataset_join != 'inner'``, otherwise ignored.
+    keep_attrs: boolean, Optional
+        Whether to copy attributes from the first argument to the output.
     exclude_dims : set, optional
         Dimensions to exclude from alignment and broadcasting. Any inputs
         coordinates along these dimensions will be dropped. Each excluded
         dimension must be a core dimension in the function signature.
-    dataset_fill_value : optional
-        Value used in place of missing variables on Dataset inputs when the
-        datasets do not share the exact same ``data_vars``. Only relevant if
-        ``join != 'inner'``.
     kwargs: dict, optional
         Optional keyword arguments passed directly on to call ``func``.
     dask_array: 'forbidden' or 'allowed', optional
@@ -664,8 +706,10 @@ def apply_ufunc(func, *args, **kwargs):
 
     signature = kwargs.pop('signature', None)
     join = kwargs.pop('join', 'inner')
+    dataset_join = kwargs.pop('dataset_join', 'inner')
+    keep_attrs = kwargs.pop('keep_attrs', False)
     exclude_dims = kwargs.pop('exclude_dims', frozenset())
-    dataset_fill_value = kwargs.pop('dataset_fill_value', None)
+    dataset_fill_value = kwargs.pop('dataset_fill_value', _DEFAULT_FILL_VALUE)
     kwargs_ = kwargs.pop('kwargs', None)
     dask_array = kwargs.pop('dask_array', 'forbidden')
     if kwargs:
@@ -697,16 +741,21 @@ def apply_ufunc(func, *args, **kwargs):
         this_apply = functools.partial(
             apply_ufunc, func, signature=signature, join=join,
             dask_array=dask_array, exclude_dims=exclude_dims,
-            dataset_fill_value=dataset_fill_value)
+            dataset_fill_value=dataset_fill_value,
+            dataset_join=dataset_join,
+            keep_attrs=keep_attrs)
         return apply_groupby_ufunc(this_apply, *args)
     elif any(is_dict_like(a) for a in args):
         return apply_dataset_ufunc(variables_ufunc, *args, signature=signature,
                                    join=join, exclude_dims=exclude_dims,
-                                   fill_value=dataset_fill_value)
+                                   fill_value=dataset_fill_value,
+                                   dataset_join=dataset_join,
+                                   keep_attrs=keep_attrs)
     elif any(isinstance(a, DataArray) for a in args):
         return apply_dataarray_ufunc(variables_ufunc, *args,
                                      signature=signature,
-                                     join=join, exclude_dims=exclude_dims)
+                                     join=join, exclude_dims=exclude_dims,
+                                     keep_attrs=keep_attrs)
     elif any(isinstance(a, Variable) for a in args):
         return variables_ufunc(*args)
     else: